Advanced/GreenIntel/requirements.txt
streamlit
folium==0.15.0
pydeck==0.8.0
numpy==1.26.3
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.1
requests==2.31.0
tensorflow>=2.16.0
scikit-learn==1.4.0
python-dotenv==1.0.0
earthdata
netCDF4==1.6.5
xarray==2023.12.0
langchain==0.1.0
langchain-community==0.0.13
chromadb==0.4.22
streamlit_folium
transformers==4.36.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
accelerate==0.25.0